+++ src/main/java/it/gov/pagopa/payhub/activities/activity/debtposition/synchronize/gpdpreload/SynchronizeInstallmentGpdPreLoadActivity.java
@@ -1,4 +1,4 @@
-package it.gov.pagopa.payhub.activities.activity.debtposition.gpdpreload;
+package it.gov.pagopa.payhub.activities.activity.debtposition.synchronize.gpdpreload;
 
 import io.temporal.activity.ActivityInterface;
 import io.temporal.activity.ActivityMethod;
