+++ src/test/java/it/gov/pagopa/payhub/activities/activity/debtposition/synchronize/aca/SynchronizeInstallmentAcaActivityTest.java
@@ -1,6 +1,6 @@
-package it.gov.pagopa.payhub.activities.activity.debtposition.gpdpreload;
+package it.gov.pagopa.payhub.activities.activity.debtposition.synchronize.aca;
 
-import it.gov.pagopa.payhub.activities.connector.pagopapayments.GpdPreLoadService;
+import it.gov.pagopa.payhub.activities.connector.pagopapayments.AcaService;
 import it.gov.pagopa.pu.debtposition.dto.generated.DebtPositionDTO;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.Test;
@@ -12,16 +12,16 @@ import org.mockito.junit.jupiter.MockitoExtension;
 import static it.gov.pagopa.payhub.activities.util.faker.DebtPositionFaker.buildDebtPositionDTO;
 
 @ExtendWith(MockitoExtension.class)
-class SynchronizeInstallmentGpdPreLoadActivityTest {
+class SynchronizeInstallmentAcaActivityTest {
 
     @Mock
-    private GpdPreLoadService gpdPreLoadServiceMock;
+    private AcaService acaServiceMock;
 
-    private SynchronizeInstallmentGpdPreLoadActivity activity;
+    private SynchronizeInstallmentAcaActivity activity;
 
     @BeforeEach
     void init() {
-        activity = new SynchronizeInstallmentGpdPreLoadActivityImpl(gpdPreLoadServiceMock);
+        activity = new SynchronizeInstallmentAcaActivityImpl(acaServiceMock);
     }
 
     @Test
@@ -29,8 +29,8 @@ class SynchronizeInstallmentGpdPreLoadActivityTest {
         String iud = "IUD";
         DebtPositionDTO debtPositionDTO = buildDebtPositionDTO();
 
-        activity.synchronizeInstallmentGpdPreLoad(debtPositionDTO, iud);
+        activity.synchronizeInstallmentAca(debtPositionDTO, iud);
 
-        Mockito.verify(gpdPreLoadServiceMock).syncInstallmentGpdPreLoad(iud, debtPositionDTO);
+        Mockito.verify(acaServiceMock).syncInstallmentAca(iud, debtPositionDTO);
     }
 }
